+++ flydra-types/src/lib.rs
@@ -35,6 +35,7 @@ pub const CALIBRATION_XML_FNAME: &str = "calibration.xml";
 pub const BRAID_METADATA_YML_FNAME: &str = "braid_metadata.yml";
 pub const README_MD_FNAME: &str = "README.md";
 pub const IMAGES_DIRNAME: &str = "images";
+pub const CAM_SETTINGS_DIRNAME: &str = "cam_settings";
 pub const RECONSTRUCT_LATENCY_HLOG_FNAME: &str = "reconstruct_latency_usec.hlog";
 pub const REPROJECTION_DIST_HLOG_FNAME: &str = "reprojection_distance_100x_pixels.hlog";
 
@@ -189,20 +190,35 @@ impl BraidCameraConfig {
 }
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
-pub struct RegisterNewCamera {
+pub struct PerCamSaveData {
+    pub ros_cam_name: RosCamName,
+    pub current_image_png: Option<Vec<u8>>,
+    pub settings_data: Option<PerCamSettingsData>,
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct PerCamSettingsData {
     /// The raw name of the camera as given by the camera itself.
     pub orig_cam_name: RawCamName,
+    /// Implementation-dependent file extension used for camera settings file.
+    pub settings_file_ext: String,
+    /// Implementation-dependent dump of camera settings at the moment it was started.
+    pub settings_on_start: String,
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct RegisterNewCamera {
     /// The name of the camera used in ROS (e.g. with '-' converted to '_').
     pub ros_cam_name: RosCamName,
     /// Location of the camera control HTTP server.
-    pub http_camserver_info: CamHttpServerInfo,
+    pub http_camserver_info: Option<CamHttpServerInfo>,
+    /// Additional data about the camera settings.
+    pub settings_data: Option<PerCamSettingsData>,
 }
 
 #[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
 pub struct UpdateImage {
-    // /// The raw name of the camera as given by the camera itself.
-    // pub orig_cam_name: RawCamName,
-    /// The name of the camera used in ROS (e.g. with '-' converted to '_').
+    /// The name of the camera.
     pub ros_cam_name: RosCamName,
     pub current_image_png: Vec<u8>,
 }