+++ image-tracker/tests/test-fmf-tracking.rs
@@ -25,6 +25,12 @@ async fn track_fmf_with_error(handle: tokio::runtime::Handle) -> fmf::FMFResult<
     #[cfg(feature = "debug-images")]
     let (_shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel::<()>();
 
+    let new_cam_data = flydra_types::RegisterNewCamera {
+        ros_cam_name: flydra_types::RosCamName::new("camera1".into()),
+        http_camserver_info: None,
+        settings_data: None,
+    };
+
     let mut ft = FlyTracker::new(
         &handle,
         &RawCamName::new("fmf".to_string()),
@@ -45,6 +51,7 @@ async fn track_fmf_with_error(handle: tokio::runtime::Handle) -> fmf::FMFResult<
         #[cfg(feature = "debug-images")]
         Some(shutdown_rx),
         None,
+        new_cam_data,
     )
     .unwrap();